requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv

langgraph>=0.0.50
langchain-core>=0.2.0

chromadb
rank_bm25
numpy

tqdm

PyYAML

google-generativeai

pandas
datasets
ragas

# parsing
pymupdf           # PyMuPDF 핵심
pdfplumber        # 표 추출 보강(옵션)
pdfminer.six      # pdfplumber 의존
python-magic
pandas

# embeddings / vector
FlagEmbedding
torch
chromadb
numpy
tqdm
scikit-learn

streamlit
requests
tavily-python

# Redis for session management and caching
redis>=4.5.0
hiredis>=2.0.0